Moved login stuff from Main.cpp to MainLogin.cpp
