+++ src/ThePier.vcxproj
@@ -184,6 +184,7 @@
     <ClCompile Include="iMessage.cpp" />
     <ClCompile Include="LoginController.cpp" />
     <ClCompile Include="Main.cpp" />
+    <ClCompile Include="MainLogin.cpp" />
     <ClCompile Include="MainTest.cpp" />
     <ClCompile Include="Member.cpp" />
     <ClCompile Include="sync.cpp" />
